# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.MenuCreateRequest:
    properties:
      icon:
        description: 图标
        maxLength: 500
        type: string
      is_enabled:
        description: 是否启用
        type: boolean
      parent_id:
        description: 父菜单ID，NULL表示主菜单
        type: integer
      sort:
        description: 排序，范围1-10
        maximum: 10
        minimum: 1
        type: integer
      title:
        description: 菜单标题
        maxLength: 100
        minLength: 1
        type: string
      type:
        description: 菜单类型
        enum:
        - aggregate
        - navigation
        - footer
        type: string
      url:
        description: 链接地址（主菜单可为空）
        maxLength: 500
        type: string
    required:
    - title
    - type
    type: object
  dto.MenuDeleteRequest:
    properties:
      children_action:
        description: 子菜单处理方式：delete-删除子菜单，upgrade-升级为主菜单
        enum:
        - delete
        - upgrade
        type: string
    type: object
  dto.MenuResponse:
    properties:
      icon:
        type: string
      id:
        type: integer
      is_enabled:
        type: boolean
      parent_id:
        type: integer
      sort:
        type: integer
      title:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
  dto.MenuTreeNode:
    properties:
      children:
        description: 子菜单列表
        items:
          $ref: '#/definitions/dto.MenuTreeNode'
        type: array
      icon:
        type: string
      id:
        type: integer
      is_enabled:
        type: boolean
      parent_id:
        type: integer
      sort:
        type: integer
      title:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
  dto.MenuUpdateRequest:
    properties:
      icon:
        description: 图标
        maxLength: 500
        type: string
      is_enabled:
        description: 是否启用
        type: boolean
      parent_id:
        description: 父菜单ID
        type: integer
      sort:
        description: 排序
        maximum: 10
        minimum: 1
        type: integer
      title:
        description: 菜单标题
        maxLength: 100
        minLength: 1
        type: string
      type:
        description: 菜单类型
        enum:
        - aggregate
        - navigation
        - footer
        type: string
      url:
        description: 链接地址
        maxLength: 500
        type: string
    required:
    - title
    - type
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact:
    email: mccsjs@example.com
    name: mccsjs
  description: 一个基于 Go 语言的现代化博客后端服务
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Blog
  version: v1
paths:
...
  /admin/menus:
    get:
      consumes:
      - application/json
      description: 按类型获取菜单树
      parameters:
      - description: '菜单类型: aggregate/navigation/footer'
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.MenuTreeNode'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 获取菜单树
      tags:
      - 菜单管理
    post:
      consumes:
      - application/json
      description: 创建新的菜单项
      parameters:
      - description: 菜单信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.MenuCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MenuResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 创建菜单
      tags:
      - 菜单管理
  /admin/menus/{id}:
    delete:
      consumes:
      - application/json
      description: 硬删除菜单项，不可恢复
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 删除选项
        in: body
        name: request
        schema:
          $ref: '#/definitions/dto.MenuDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除菜单
      tags:
      - 菜单管理
    get:
      consumes:
      - application/json
      description: 获取菜单项的详细信息
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MenuResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取菜单详情
      tags:
      - 菜单管理
    put:
      consumes:
      - application/json
      description: 更新菜单项信息
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 菜单信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.MenuUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MenuResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 更新菜单
      tags:
      - 菜单管理
...
  /menus:
    get:
      consumes:
      - application/json
      parameters:
      - description: '菜单类型: aggregate/navigation/footer (不传则返回所有类型)'
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.MenuTreeNode'
                  type: array
              type: object
      summary: 获取菜单树
      tags:
      - 菜单
...
securityDefinitions:
  BearerAuth:
    description: 在请求头中添加 Bearer Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"